Factor out simple web server
